# Performance parity for 32-bit-precision in FP16 ARM matrix-vector kernel using FMLAL instruction